Updated Finagle and Scala versions
Problem

The project was a couple of Finagle versions behind, and there have been some
changes to the Stack API in the meantime.

Solution

I've had to make some changes to the public API; in particular there's no more
`ZookeeperClient` or `ZookeeperStackClient`. I'm not sure the implementation of
`Zookeeper.newClient` is ideal, but all tests pass.

Result

Updated to Finagle 6.24.0, which also means we can cross-build for Scala 2.11.